feat(store): refactor game store methods and computed properties into separate files
